use long value for socket to avoid lost connections.
git-svn-id: http://bluecove.googlecode.com/svn/trunk@575 3cac39a0-5d1b-0410-8711-c1891c4e0f47

+++ src/main/java/com/intel/bluetooth/BluetoothPeer.java
@@ -107,37 +107,37 @@ public class BluetoothPeer {
 	/*
 	 * socket operations
 	 */
-	public native int socket(boolean authenticate, boolean encrypt) throws IOException;
+	public native long socket(boolean authenticate, boolean encrypt) throws IOException;
 
-	public native long getsockaddress(int socket) throws IOException;
+	public native long getsockaddress(long socket) throws IOException;
 	
-	public native void storesockopt(int socket);
+	public native void storesockopt(long socket);
 
-	public native int getsockchannel(int socket) throws IOException;
+	public native int getsockchannel(long socket) throws IOException;
 
-	public native void connect(int socket, long address, int channel) throws IOException;
+	public native void connect(long socket, long address, int channel) throws IOException;
 
-	public native void bind(int socket) throws IOException;
+	public native void bind(long socket) throws IOException;
 	
-	public native void listen(int socket) throws IOException;
+	public native void listen(long socket) throws IOException;
 
-	public native int accept(int socket) throws IOException;
+	public native long accept(long socket) throws IOException;
 
-	public native long recvAvailable(int socket) throws IOException;
+	public native int recvAvailable(long socket) throws IOException;
 	
-	public native int recv(int socket) throws IOException;
+	public native int recv(long socket) throws IOException;
 
-	public native int recv(int socket, byte[] b, int off, int len) throws IOException;
+	public native int recv(long socket, byte[] b, int off, int len) throws IOException;
 
-	public native void send(int socket, int b) throws IOException;
+	public native void send(long socket, int b) throws IOException;
 
-	public native void send(int socket, byte[] b, int off, int len) throws IOException;
+	public native void send(long socket, byte[] b, int off, int len) throws IOException;
 
-	public native void close(int socket) throws IOException;
+	public native void close(long socket) throws IOException;
 
 	public native String getpeername(long address) throws IOException;
 
-	public native long getpeeraddress(int socket) throws IOException;
+	public native long getpeeraddress(long socket) throws IOException;
 
 	public native String getradioname(long address);
